feat(ctb): Add new move type to FDG for OR counters
Adds a potential new move type to the `FaultDisputeGame` that allows for
a participant to reveal the preimage of the claimed output root to
display to the dispute game that the claimed L2 block number does not
match up with the block number that the block header within the output
root commits to.

The root output can be challenged with the new special move type iff:
1. The passed `OutputRootProof` hashes to equal the claimed output root.
1. The passed Header RLP hashes to equal the block hash within the
`OutputRootProof` above.
1. The claimed block number in the dispute game does not equal the block
number that the output root commits to.

If there is a successful challenge with the new move type, that claim
itself is inserted as a special subgame of the root claim at position
`0`. Notably, _position `0` is an invalid position, and cannot be
traversed to under normal circumstances_.

As this move type reaches its terminal case as soon as it goes through,
it may never be countered. As such, a new rule has been added to the
`move` function that disallows any moves against a claim with position
`0`.